Implement a queue on the client and server side.
This is a large changeset to implement a client and
server side jobs queue. We want the requests to be
asynchronous so we really need to queue on the client
side. To achieve this the queue is now a generic
package. The proper build procedure is a generic
formal procedure. On the client side the Build.Run
is used, on the server side a requests is sent to
every client that are registered to run the give
project/scenario.