Add support for regexp attribute on cmd nodes.
The regexp is used to retreive only part of the output
of the given command. As an example this is used in
the Subversion support to get the current repository
version out of the svn info output.

Add a test for this new attribute. At the same time
enhance the client/server to do some automatic status
checking before exiting. We check that some files have
been retreived and built.