Do not return the 404 error code in case of error.
When an error is detected and we return an error message
we do not want to use the 404 error code which must be
used only when a page is not found. For now use the
standard 202 code.

Move also an exception handler to ensure that the exception
raise in a declare part is properly handled.