Add support for build latency.
A build was always done as soon as the request was received. This
is not right when multiple commits are related. In this case we do
not want to start the build immediately but let some latency. If a
new build request is received for the same project and scenario the
registered build is move in time to allow the same latency. This
avoids build failures when related commits are done.