Add support for filters.
The initial implementation of the SCM file_updated specifc node
has been removed. A more general support for parsing command
output is now supported through filters. This is the first step
forward full filter support. A later feature will be to control
the action execution depending on filters output. The final
goal is to be able to skip a build if all updated files are
ignored (not part of the application, like readme.txt for
example).